refactor: file-separated tests to avoid having one long mega-file
