+++ tests/basic/basic.py
@@ -1,0 +1,8 @@
+import requests
+import pytest
+
+def test_404():
+    url = "http://localhost:8000/does_not_exist"
+    response = requests.get(url)
+    assert response.status_code == 404
+    assert "error" in response.json()
